Tiled Software Raster Overhaul + Device/D3D11 start
Rasterizer:
- Use 32 bit bitmasks instead of 64 bit (so that it can be properly supported on GPU)
- Use SSE in the rasterizer (todo: AVX/AVX2?)
- 16 bit color mode working (mostly)
- Reworked the vbuffer copy/fill etc to be more in line with original, fixed some bugs
- Disabled various old GPU draw functions
- Improved the final window blit

Device handling:
- Device menu now works correctly, can swap devices on the fly
- New restart mode restarts the game if the device is changed instead of doing it whenever the game/menu swap happens (which was why the original game had such a hard time swapping between menu and game)
- All 2D/3D resources can be loaded in HW now thanks to the device change approach

Backends:
- D3D11 backend started-
- Can be selected from the display menu
- The vbuffer stuff is working at least in 8 bit mode
- Skeleton for the rasterizer is somewhat in place but very unfinished
- Debating replacing the shared vmem stuff (which would allow for all formats to be supported) and software compute raster stuff with just classic textures and draw queues like the GL version as that might be more portable, but doing the 8 bit mode would be more difficult

Compiler/CMAKE:
- Now compile HLSL files automatically and place their compiled blobs into the resource director for engine embedding

+++ src/Engine/rdCanvas.c
@@ -52,10 +52,10 @@ int rdCanvas_NewEntry(rdCanvas *canvas, int bIdk, stdVBuffer *vbuf, stdVBuffer *
 	canvas->tileHeight = ((height + RDCACHE_FINE_TILE_SIZE - 1) / RDCACHE_FINE_TILE_SIZE);
 	canvas->tileCount = canvas->tileWidth * canvas->tileHeight;
 
-	canvas->coarseTileBits = rdroid_pHS->alloc(canvas->coarseTileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint64_t));
-	canvas->tileBits = rdroid_pHS->alloc(canvas->tileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint64_t));
-	memset(canvas->coarseTileBits, 0, canvas->coarseTileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint64_t));
-	memset(canvas->tileBits, 0, canvas->tileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint64_t));
+	canvas->coarseTileBits = rdroid_pHS->alloc(canvas->coarseTileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint32_t));
+	canvas->tileBits = rdroid_pHS->alloc(canvas->tileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint32_t));
+	memset(canvas->coarseTileBits, 0, canvas->coarseTileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint32_t));
+	memset(canvas->tileBits, 0, canvas->tileCount * RDCACHE_TILE_BINNING_STRIDE * sizeof(uint32_t));
 #endif
     return 1;
 }
